Route all tests through a central JerseyTest
This patch creates a parent test class that manages global test
configuration values.
