+++ kangaroo-common/src/test/java/net/krotscheck/kangaroo/common/status/StatusServiceTest.java
@@ -18,14 +18,12 @@
 
 package net.krotscheck.kangaroo.common.status;
 
-import net.krotscheck.kangaroo.common.config.ConfigurationFeature;
 import org.apache.http.HttpStatus;
 import org.glassfish.jersey.server.ResourceConfig;
-import org.glassfish.jersey.test.JerseyTest;
+import net.krotscheck.kangaroo.test.KangarooJerseyTest;
 import org.junit.Assert;
 import org.junit.Test;
 
-import javax.ws.rs.core.Application;
 import javax.ws.rs.core.Response;
 
 /**
@@ -33,7 +31,7 @@ import javax.ws.rs.core.Response;
  *
  * @author Michael Krotscheck
  */
-public final class StatusServiceTest extends JerseyTest {
+public final class StatusServiceTest extends KangarooJerseyTest {
 
     /**
      * Setup the test application.
@@ -41,7 +39,7 @@ public final class StatusServiceTest extends JerseyTest {
      * @return A configured application for this test harness.
      */
     @Override
-    protected Application configure() {
+    protected ResourceConfig createApplication() {
         ResourceConfig a = new ResourceConfig();
         a.register(StatusService.class);
         return a;
